pytests/pyproject.toml
[project]
name = "pytests"
version = "0.1.0"
requires-python = "==3.13.*"
dependencies = [
    "pandas>=2.3.2",
    "psycopg>=3.2.10",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "tabulate>=0.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
